pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bosch-biometric-mlops"
version = "0.1.0"
description = "Multimodal biometric recognition system with production MLOps practices"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "ray>=2.9",
    "pyarrow>=14.0",
    "Pillow>=10.0",
    "numpy>=1.26",
    "scikit-learn>=1.3",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.6",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["biometric"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
packages = ["biometric"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "ray.*",
    "hydra.*",
    "omegaconf.*",
    "torchvision.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring full dataset",
]
